feat: added destroying of part.file if the node request ends with an error event (#319)
If the client cancels the http request (e.g. just closes the socket),
the underlying file stream will not be closed with an error but
the request handling will hang.

The tests saves the stream of a file part into a Noop write stream
through pipeline. If the client cancels the request, the pipeline
call will not return. With the fix, we keep track of the file that
is currently being processed and in the case of an error event on the
node request, this file will be destroyed.